fix hourly temperature to be double

+++ src/main/kotlin/logic/model/HourlyTemperature.kt
@@ -1,6 +1,6 @@
 package logic.model
 
 data class HourlyTemperature(
-    val temperature: Float,
+    val temperature: Double,
     val hour: Int,
 )
